Add bottom tab navigation with placeholder screens
