+++ package.json
@@ -14,7 +14,11 @@
     "expo-status-bar": "~2.2.3",
     "react": "19.0.0",
     "react-native": "0.79.2",
-    "react-native-reanimated": "~3.17.4"
+    "react-native-reanimated": "~3.17.4",
+    "@react-navigation/native": "^6.1.8",
+    "@react-navigation/bottom-tabs": "^6.5.8",
+    "react-native-safe-area-context": "^4.8.0",
+    "react-native-screens": "^3.25.0"
   },
   "devDependencies": {
     "@babel/core": "^7.27.1",
