+++ screens/TimetableScreen.tsx
@@ -1,0 +1,17 @@
+import { View, Text, StyleSheet } from 'react-native';
+
+export default function TimetableScreen() {
+  return (
+    <View style={styles.container}>
+      <Text>Timetable Screen</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
